requirements.txt
alembic==1.3.2
aniso8601==8.0.0
Click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==3.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.1.0
MarkupSafe==2.1.1
pandas
psycopg2-binary==2.9.2
psycopg2==2.8.6
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.3
six==1.13.0
Werkzeug==2.2.2
sqlalchemycollector==1.3.17